% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqc_main.R
\name{cv.pqc}
\alias{cv.pqc}
\title{Cross Validation for PQC}
\usage{
cv.pqc(data, projDimRange, tau, lambdaRange, muEst = TRUE,
  epsilon = 1/log10(nrow(data)), shareNA = 0.1, iterTol = 200,
  convTol = NA, progBar = TRUE, doPar = TRUE, doSeq = FALSE)
}
\arguments{
\item{data}{data matrix with rows as data entries and columns as variables}

\item{projDimRange}{# range of desired principal components}

\item{tau}{asymmetry parameter (b'w 0 and 1)}

\item{lambdaRange}{regulization parameter}

\item{muEst}{calculate a constant mu}

\item{epsilon}{approxiamtion parameter}

\item{shareNA}{size of holdout (0<shareNA<1)}

\item{iterTol}{no. of max iterataions}

\item{convTol}{set algorithm to stop if weigths did not change for convTol no. of consecutive iterations (deactivated for tau=0.5)}

\item{progBar}{optional progressbar}

\item{doPar}{use parallel backend (*nix systems recommended),}

\item{doSeq}{run via sequential optimization}

\item{preOut}{continues based on previous output (if doSeq is TRUE, the components of a previos PCA method are sufficient)}
}
\value{
list containing pqcomp() results
}
\description{
Wrapper for pqcomp(): Select ranges for rank and lambda.
Results are calculated on a random hold out on the supplied dataset.
}
\references{
Madeleine Udell et al. (2016), "Generalized Low Rank Models".
}
