% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqc_main.R
\name{cv.scores}
\alias{cv.scores}
\title{Calculate cross validation scores via quantile check loss on the result of cv.pqc().}
\usage{
cv.scores(cv.result, tau)
}
\arguments{
\item{cv.result}{list output of cv.pqc()}

\item{tau}{asymmetry parameter (0<tau<1)}
}
\value{
list containing pqcomp() results
}
\description{
Calculate cross validation scores via quantile check loss on the result of cv.pqc().
}
