% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqc_main.R
\name{pqcomp}
\alias{pqcomp}
\title{Main quantile components algorithm}
\usage{
pqcomp(data, projDim, tau, lambda, muEst = TRUE,
  epsilon = 1/log10(nrow(data)), iterTol = 200, convTol = NA,
  preOut = NA, progBar = TRUE, doPar = TRUE, doSeq = FALSE,
  randomSeed = NA)
}
\arguments{
\item{data}{data matrix with rows as data entries and columns as variables}

\item{projDim}{no. of desired principal components}

\item{tau}{regulization parameter}

\item{lambda}{regulization parameter}

\item{muEst}{calculate a constant mu}

\item{epsilon}{approxiamtion parameter}

\item{iterTol}{no. of max iterataions}

\item{convTol}{set algorithm to stop if weigths did not change for convTol no. of consecutive iterations (deactivated for tau=0.5)}

\item{preOut}{continues based on previous output (if doSeq is TRUE, the components of a previos PCA method are sufficient)}

\item{progBar}{optional progressbar}

\item{doPar}{use parallel backend (*nix systems recommended),}

\item{doSeq}{run via sequential optimization}

\item{randomSeed}{select a random seed as in set.seed for fixed initialization}
}
\value{
list object containing components and loadings
}
\description{
Main quantile components algorithm
}
